Add files via upload
In this exercise, you are to re-write the recursive implementation of NchooseK() you wrote for Lab 3, with a minor change to the input format: the arguments n and k will be on separate lines in the input (because MARS can only handle one number per input line). The definition of the function is repeated here for convenience:

        NchooseK(n, 0) = 1
        NchooseK(n, n) = 1
        NchooseK(n, k) = NchooseK(n-1, k-1) + NchooseK(n-1, k)
All other assumptions from Lab 3 still hold. The implementation of the NchooseK function must be recursive. A non-recursive implementation of the function will receive zero credit. You can assume that the function will be called only with an argument small enough so that the result does not overflow, i.e., fits within a 32-bit integer.

First implement this function in C (or copy your implementation from Lab 3 if it was correct, and modify the part that reads the input). Note the following:

A starter file is provided: ex1.c.
The function should have the following name and type: int NchooseK(int n, int k)
The main() procedure repeatedly reads two lines of input, calls NchooseK() to compute NchooseK, and finally uses printf("%d\n",...) to print the NchooseK value.
These steps are repeated until a 0 value is read by main as the first argument, and then the program terminates.
Name the file with your C code ex1.c. Test it using your own inputs, and also the sample inputs provided.

Assembly version

Once your C code is working correctly, write the equivalent MIPS assembly program, using the provided starter file (ex1.asm). Since you are implementing a procedure that is recursive, you will need to appropriately manage the stack. That is the main challenge of this exercise